requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
fastapi==0.68.1
fastapi_chameleon==0.1.12
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.2
psycopg2-binary==2.9.1
pydantic==1.8.2
PyJWT==2.1.0
python-decouple==3.5
python-http-client==3.3.3
pytz==2021.3
requests==2.26.0
sendgrid==6.8.2
SQLAlchemy==1.4.25
sqlalchemy2-stubs==0.0.2a17
sqlmodel==0.0.4
starkbank-ecdsa==1.1.1
starlette==0.14.2
twilio==7.0.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
